refactor: refactor share crossfade dialog
